Creating new file where all classes and functionality concern ACCU Weather API conection are stored. Creating ACCUWeatherAPI class which response for all api connection and two previous classes only converts data to dictionary
